refactor: 피드백 반영 후 구조 변경
