+++ src/main/java/edu/example/wayfarer/api_controller/MainController.java
@@ -52,9 +52,9 @@ public class MainController {
         Member currentUser = securityUtil.getCurrentUser();
 //        memberRoomRequestDTO.email() = currentUser.getEmail();
         MemberRoomRequestDTO updatedDTO = new MemberRoomRequestDTO(
-                currentUser.getEmail(),
                 memberRoomRequestDTO.roomId(),
-                memberRoomRequestDTO.roomCode()
+                memberRoomRequestDTO.roomCode(),
+                currentUser.getEmail()
         );
         memberRoomService.create(updatedDTO);
         return ResponseEntity.ok(Map.of("message", "방에 입장했습니다."));
